prototype/back-end/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
werkzeug==3.0.1
marshmallow==3.20.1

